Added workaround for false positives from static analysis
